feat(ui): Added otages import step. Extracted common file-upload related logic into FileUploadViewModel
