+++ Alerting.ML.App/App.axaml
@@ -87,6 +87,12 @@
             <Setter Property="BorderThickness" Value="1"/>
         </Style>
 
+        <Style Selector="Border.note">
+            <Setter Property="Background" Value="{StaticResource Color.NoteBorder}" />
+            <Setter Property="Padding" Value="12"/>
+            <Setter Property="CornerRadius" Value="8"/>
+        </Style>
+
         <Style Selector="Border.success">
             <Setter Property="Background" Value="{StaticResource Color.SuccessPanel}" />
             <Setter Property="BorderBrush" Value="{StaticResource Color.SuccessPanelBorder}"/>
