+++ Alerting.ML.App/Components/TrainingCreation/Csv/TrainingCreationCsvSecondStepViewModel.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Reactive;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
+using Alerting.ML.App.Components.TrainingCreation.FileUpload;
 
 namespace Alerting.ML.App.Components.TrainingCreation.Csv;
 
@@ -18,29 +19,16 @@ using Alerting.ML.App.Model.Enums;
 using Alerting.ML.App.ViewModels;
 using Avalonia.Controls;
 
-public class TrainingCreationCsvSecondStepViewModel : ViewModelBase, ITrainingCreationStepViewModel
+public class TrainingCreationCsvSecondStepViewModel : FileUploadViewModel, ITrainingCreationStepViewModel
 {
     private readonly TrainingBuilder builder;
     public string? UrlPathSegment => "csv";
-
-    public string SelectedFilePath
-    {
-        get;
-        set => this.RaiseAndSetIfChanged(ref field, value);
-    }
-
-    public string SelectedFileName => System.IO.Path.GetFileName(SelectedFilePath);
-
+    
     public bool IsAzureScheduledQueryRuleSelected
     {
         get;
         set => this.RaiseAndSetIfChanged(ref field, value);
     }
-
-    public bool IsFileSelected => !string.IsNullOrWhiteSpace(SelectedFilePath);
-
-    public string UploadTitle => IsFileSelected ? SelectedFileName : "Upload CSV File";
-    public string UploadSubTitle => IsFileSelected ? "Click to change file" : "Drag and drop or click to browse";
 
     public IScreen HostScreen { get; }
 
@@ -58,49 +46,12 @@ public class TrainingCreationCsvSecondStepViewModel : ViewModelBase, ITrainingCr
 
     public TrainingCreationStep CurrentStep => TrainingCreationStep.Step2;
 
-    public ReactiveCommand<IEnumerable<IStorageItem>, Unit> FileDroppedCommand { get; }
-    public ReactiveCommand<Visual, Unit> PickFileCommand { get; }
+    protected override string Title => "Upload CSV File";
 
     public TrainingCreationCsvSecondStepViewModel(IScreen hostScreen, TrainingBuilder builder)
     {
         this.builder = builder;
         HostScreen = hostScreen;
-        FileDroppedCommand = ReactiveCommand.Create<IEnumerable<IStorageItem>>(FileDropped);
-        PickFileCommand = ReactiveCommand.CreateFromTask<Visual>(PickFile);
-        this.WhenAnyValue(model => model.SelectedFilePath)
-            .Subscribe(value =>
-            {
-                this.RaisePropertyChanged(nameof(IsFileSelected));
-                this.RaisePropertyChanged(nameof(UploadTitle));
-                this.RaisePropertyChanged(nameof(UploadSubTitle));
-                this.RaisePropertyChanged(nameof(SelectedFileName));
-            });
-    }
-
-    private async Task PickFile(Visual control)
-    {
-        var topLevel = TopLevel.GetTopLevel(control);
-
-        var files = await topLevel!.StorageProvider.OpenFilePickerAsync(
-            new FilePickerOpenOptions
-            {
-                Title = "Select a CSV file",
-                AllowMultiple = false,
-                FileTypeFilter = new[]
-                {
-                    new FilePickerFileType("CSV Files")
-                    {
-                        Patterns = new[] { "*.csv" }
-                    }
-                }
-            });
-
-        SelectedFilePath = files.First().Path.ToString();
-    }
-
-    private void FileDropped(IEnumerable<IStorageItem> arg)
-    {
-        SelectedFilePath = arg.First().Path.ToString();
     }
 }
